Increase the number of plots the plugin can handle. Update Plot Limit to 2^(31)-1
Conflicts:
	src/main/java/com/worldcretornica/plotme_core/commands/CmdAuto.java
	src/main/java/com/worldcretornica/plotme_core/commands/CmdBuy.java
	src/main/java/com/worldcretornica/plotme_core/commands/CmdClaim.java

+++ src/main/java/com/worldcretornica/plotme_core/SqlManager.java
@@ -367,7 +367,7 @@ public class SqlManager {
                     Statement slMetadata = sqliteconn.createStatement();
                     ResultSet setMetadata = null;
 
-                    short size = 0;
+                    int size = 0;
                     while (setPlots.next()) {
                         int idX = setPlots.getInt("idX");
                         int idZ = setPlots.getInt("idZ");
@@ -1172,11 +1172,11 @@ public class SqlManager {
      * @param world plotworld to check
      * @return number of plots in the world
      */
-    public short getPlotCount(String world) {
+    public int getPlotCount(String world) {
+
         PreparedStatement ps = null;
         ResultSet setNbPlots = null;
-        short nbplots = 0;
-
+        int nbplots = 0;
         try {
             Connection conn = getConnection();
 
@@ -1186,7 +1186,7 @@ public class SqlManager {
             setNbPlots = ps.executeQuery();
 
             if (setNbPlots.next()) {
-                nbplots = setNbPlots.getShort(1);
+                nbplots = setNbPlots.getInt(1);
             }
         } catch (SQLException ex) {
             plugin.getLogger().severe("PlotCount Exception :");
@@ -1207,10 +1207,10 @@ public class SqlManager {
         return nbplots;
     }
 
-    public short getPlotCount(String world, UUID ownerId, String owner) {
+    public int getPlotCount(String world, UUID ownerId, String owner) {
         PreparedStatement ps = null;
         ResultSet setNbPlots = null;
-        short nbplots = 0;
+        int nbplots = 0;
 
         try {
             Connection conn = getConnection();
@@ -1228,7 +1228,7 @@ public class SqlManager {
             setNbPlots = ps.executeQuery();
 
             if (setNbPlots.next()) {
-                nbplots = setNbPlots.getShort(1);
+                nbplots = setNbPlots.getInt(1);
             }
         } catch (SQLException ex) {
             plugin.getLogger().severe("PlotCount Exception :");
@@ -1249,10 +1249,10 @@ public class SqlManager {
         return nbplots;
     }
 
-    public short getFinishedPlotCount(String world) {
+    public int getFinishedPlotCount(String world) {
         PreparedStatement ps = null;
         ResultSet setNbPlots = null;
-        short nbplots = 0;
+        int nbplots = 0;
 
         try {
             Connection conn = getConnection();
@@ -1263,7 +1263,7 @@ public class SqlManager {
             setNbPlots = ps.executeQuery();
 
             if (setNbPlots.next()) {
-                nbplots = setNbPlots.getShort(1);
+                nbplots = setNbPlots.getInt(1);
             }
         } catch (SQLException ex) {
             plugin.getLogger().severe("FinishedPlotCount Exception :");
@@ -1284,10 +1284,10 @@ public class SqlManager {
         return nbplots;
     }
 
-    public short getExpiredPlotCount(String world) {
+    public int getExpiredPlotCount(String world) {
         PreparedStatement ps = null;
         ResultSet setNbPlots = null;
-        short nbplots = 0;
+        int nbplots = 0;
 
         try {
             Connection conn = getConnection();
@@ -1303,7 +1303,7 @@ public class SqlManager {
             setNbPlots = ps.executeQuery();
 
             if (setNbPlots.next()) {
-                nbplots = setNbPlots.getShort(1);
+                nbplots = setNbPlots.getInt(1);
             }
         } catch (SQLException ex) {
             plugin.getLogger().severe("ExpiredPlotCount Exception :");
